Get 100 log lines
The fetched number of log lines (10) was too low s.t. when a container
took some time to boot the 404 messages from the request already took up
too much space in the logs, and the `Ready.` token could not be
detected.
This commit increases the number of log lines to 100, so that this error
should be alleviated, although it is of course not an optimal solution.

It might make sense too simply fetch all logs (with a -1 paramter), and
let the user overwrite this value.